backend/pyproject.toml
[project]
name = "project-management-backend"
version = "0.1.0"
description = "Professional Multi-Tenant Project Management System Backend API"
authors = [{ name = "Sharmeen Asif" }]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.110.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "uvicorn[standard]>=0.25.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "bcrypt>=4.0.0",
    "aiofiles>=23.2.0",
    "aiosmtplib>=3.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pyright>=1.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportMissingTypeStubs = false

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html"
